+++ extensions/iceworks-material-helper/src/extension.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { connectService, getHtmlForWebview } from '@iceworks/vscode-webview/lib/vscode';
-import { initExtension } from '@iceworks/common-service';
+import { initExtension, registerCommand } from '@iceworks/common-service';
 import { Recorder, recordDAU } from '@iceworks/recorder';
 import services from './services/index';
 import propsAutoComplete from './propsAutoComplete';
@@ -60,7 +60,7 @@ export function activate(context: vscode.ExtensionContext) {
     }
   }
   subscriptions.push(
-    vscode.commands.registerCommand('iceworks-material-helper.start', function () {
+    registerCommand('iceworks-material-helper.start', function () {
       activeWebview();
     })
   );