+++ extensions/iceworks-ui-builder/web/src/pages/PageGenerater/index.tsx
@@ -51,7 +51,7 @@ const Home = () => {
 
   function validateData({ blocks }) {
     if (!blocks.length) {
-      return intl.formatMessage({ id: 'web.iceworksPageBuilder.Home.enterPageName' });
+      return intl.formatMessage({ id: 'web.iceworksUIBuilder.pageGenerater.enterPageName' });
     }
     return '';
   }
@@ -141,14 +141,16 @@ const Home = () => {
 
     setIsCreating(false);
     setVisible(false);
-    Notification.success({ content: intl.formatMessage({ id: 'web.iceworksPageBuilder.Home.successCreatePage' }) });
+    Notification.success({
+      content: intl.formatMessage({ id: 'web.iceworksUIBuilder.pageGenerater.successCreatePage' }),
+    });
     resetData();
   }
 
   return (
     <div className={styles.wrap}>
       <div className={styles.label}>
-        <FormattedMessage id="web.iceworksPageBuilder.Home.chooseBlock" />
+        <FormattedMessage id="web.iceworksUIBuilder.pageGenerater.chooseBlock" />
       </div>
       <div className={styles.field}>
         <Row gutter={24} className={styles.row}>
@@ -181,7 +183,7 @@ const Home = () => {
       </div>
       <div className={styles.opts}>
         <Button type="primary" size="medium" onClick={handleCreate}>
-          <FormattedMessage id="web.iceworksPageBuilder.Home.createPage" />
+          <FormattedMessage id="web.iceworksUIBuilder.pageGenerater.createPage" />
         </Button>
       </div>
       <RouterDetailForm