+++ extensions/iceworks-app/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 0.4.3
+
+- fix: build error
+
+## 0.4.2
+
+- fix: can't get registerCommand method callback args
+
+## 0.4.1
+
+- fix: build error
+
+## 0.4.0
+
+- refactor: auto start status bar and editor title 
+- feat: delete page & component
+
 ## 0.3.0
 
 - feat: support en-US