+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 0.3.4
+
+- fix: build error
+
+## 0.3.3
+
+- fix: can't get registerCommand method callback args
+
+## 0.3.2
+
+- fix: build error
+
+## 0.3.1
+
+- chore: change data record
+
 ## 0.3.0
 
 - feat: support en-US