+++ extensions/iceworks-ui-builder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 0.1.3
+
+- fix: build error
+
+## 0.1.2
+
+- fix: can't get registerCommand method callback args
+
+## 0.1.1
+
+- fix: build error
+
 ## 0.1.0
 
 - feat: Create Component by Materials